+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * [BUGFIX] Fix regexp parsing panic for regexp label matchers with start/end quantifiers. #1883
 * [BUGFIX] Ingester: fixed deceiving error log "failed to update cached shipped blocks after shipper initialisation", occurring for each new tenant in the ingester. #1893
 * [BUGFIX] Ring: fix bug where instances may appear unhealthy in the hash ring web UI even though they are not. #1933
+* [BUGFIX] API: gzip is now enforced when identity encoding is explicitly rejected. #1864
 
 ### Mixin